% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vvs_mco_lamda.R
\name{dico_v}
\alias{dico_v}
\title{Obtenir le dictionnaire des colonnes}
\usage{
dico_v(annee, type_v, tolower_names = TRUE)
}
\description{
Obtenir le dictionnaire des colonnes
}
\examples{
# dico_v(2017, 'valo')
# dico_v(2019, 'valo.lamda')
}
