% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vvs_mco_lamda.R
\name{format_v}
\alias{format_v}
\title{Obtenir le format pour un fichier visual}
\usage{
format_v(path, file)
}
\description{
Obtenir le format pour un fichier visual
}
\examples{
# path = "~/Documents/data/lmd/vvslamda"
# file = "750712184.2019.1.valo.lamda.txt"
# format_v(path, file)
}
