% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vvs_mco_lamda.R
\name{import_v}
\alias{import_v}
\title{Importer un fichier visual pmsi}
\usage{
import_v(path, file, tolower_names = TRUE, ...)
}
\description{
Importer un fichier visual pmsi
}
\examples{

# path = "~/Documents/data/lmd/vvslamda"
# file = "750712184.2019.1.valo.lamda.txt"
# import_v(path, file)

# path = "~/Documents/data/vvs"
# file = "750712184.2018.12.valo.txt"
# import_v(path, file)
}
