% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vvs_mco_lamda.R
\name{ouvrir_pdf_v}
\alias{ouvrir_pdf_v}
\title{Ouvrir le pdf de documentation du logiciel ATIH}
\usage{
ouvrir_pdf_v(annee, type_v)
}
\description{
Ouvrir le pdf de documentation du logiciel ATIH
}
\examples{
# pdf_v(2017, 'valo')
# pdf_v(2019, 'valo.lamda')
}
